requirements/dev.txt
-r base.txt
django-debug-toolbar
mypy
django-stubs
black
flake8
isort
